pyproject.toml
[project]
name = "streaming-overview-tui"
version = "0.1.0"
description = "Search movies and TV shows across your streaming services"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pillow>=11.0.0",
    "platformdirs>=4.5.1",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-levenshtein>=0.27.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "sqlmodel>=0.0.31",
    "textual>=7.3.0",
    "thefuzz>=0.22.1",
]

[project.scripts]
streaming-tui = "streaming_overview_tui.run:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
